Refactor placement of useFeeLiquidityMap hooks
